+++ scripts/lichessTools-init.js
@@ -94,6 +94,7 @@ lichessTools.loadTool(LiChessTools.Tools.ForkBehavior);
 lichessTools.loadTool(LiChessTools.Tools.BlogAutosave);
 lichessTools.loadTool(LiChessTools.Tools.AutoAnalyseGame);
 lichessTools.loadTool(LiChessTools.Tools.ActiveIcon);
+lichessTools.loadTool(LiChessTools.Tools.InboxChatImprovements);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);