+++ scripts/service_worker.js
@@ -38,6 +38,27 @@ importScripts('obs-websocket.js')
       if (!obs) return;
       const response = await obs.call('GetSceneList').catch(e=>{ console.debug('Could not get scenes',data); });
       return { sceneNames: response.scenes.map(s=>s.sceneName) };
+    },
+    pasteBuffer:async (data)=>{
+      const buffer = data?.options?.buffer;
+      if (!buffer) return;
+      const formData = new FormData();
+      formData.append('image', buffer);
+      formData.append('type', 'base64');
+      formData.append('title', 'LiChess Tools image');
+      formData.append('description', 'from Inbox chat paste in Lichess.org');
+      const response=await fetch(
+        'https://api.imgur.com/3/image',
+        {
+          method: 'POST',
+          headers: new Headers({
+            Authorization: 'Client-ID dd38da6dc32098d',
+          }),
+          body: formData
+        });
+      if (!response.ok) throw new Error('Imgur upload failed with status '+response.status);
+      const responseData=await response.json();
+      return { link: responseData.data?.link };
     }
   };